backend/pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = "Backend to the workflow builder project."
authors = ["kris927b <kristian.smed.noergaard@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.112.2"
sqlalchemy = "^2.0.32"
alembic = "^1.13.2"
uvicorn = "^0.30.6"
passlib = "^1.7.4"
pydantic = "^2.8.2"
python-jose = {extras = ["pycryptodome"], version = "^3.3.0"}
pydantic-settings = "^2.4.0"


[tool.poetry.group.dev.dependencies]
black = "^24.8.0"
isort = "^5.13.2"
pytest = "^8.3.2"
httpx = "^0.27.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
